+++ sdk/communication/communication-rooms/samples/v1/javascript/package.json
@@ -1,14 +1,10 @@
 {
-  "name": "@azure-samples/communication-rooms-ts-beta",
+  "name": "@azure-samples/communication-rooms-js",
   "private": true,
   "version": "1.0.0",
-  "description": "Azure Communication Services - Rooms client library samples for TypeScript (Beta)",
+  "description": "Azure Communication Services - Rooms client library samples for JavaScript",
   "engines": {
     "node": ">=12.0.0"
-  },
-  "scripts": {
-    "build": "tsc",
-    "prebuild": "rimraf dist/"
   },
   "repository": {
     "type": "git",
@@ -32,14 +28,9 @@
   },
   "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/communication/communication-rooms",
   "dependencies": {
-    "@azure/communication-rooms": "next",
+    "@azure/communication-rooms": "latest",
     "dotenv": "latest",
     "@azure/communication-identity": "^1.1.0-beta.2",
     "@azure/communication-common": "~2.1.0"
-  },
-  "devDependencies": {
-    "@types/node": "^12.0.0",
-    "typescript": "~4.4.0",
-    "rimraf": "latest"
   }
 }
