+++ sdk/communication/communication-rooms/samples/v1/javascript/sample.env
