+++ sdk/communication/communication-rooms/samples/v1/typescript/package.json
@@ -1,10 +1,14 @@
 {
-  "name": "@azure-samples/communication-rooms-js-beta",
+  "name": "@azure-samples/communication-rooms-ts",
   "private": true,
   "version": "1.0.0",
-  "description": "Azure Communication Services - Rooms client library samples for JavaScript (Beta)",
+  "description": "Azure Communication Services - Rooms client library samples for TypeScript",
   "engines": {
     "node": ">=12.0.0"
+  },
+  "scripts": {
+    "build": "tsc",
+    "prebuild": "rimraf dist/"
   },
   "repository": {
     "type": "git",
@@ -28,9 +32,14 @@
   },
   "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/communication/communication-rooms",
   "dependencies": {
-    "@azure/communication-rooms": "next",
+    "@azure/communication-rooms": "latest",
     "dotenv": "latest",
     "@azure/communication-identity": "^1.1.0-beta.2",
     "@azure/communication-common": "~2.1.0"
+  },
+  "devDependencies": {
+    "@types/node": "^12.0.0",
+    "typescript": "~4.4.0",
+    "rimraf": "latest"
   }
 }
