+++ sdk/communication/communication-rooms/samples/v1/typescript/sample.env
