+++ sdk/communication/communication-rooms/samples/v1/typescript/src/participantOperations.ts
@@ -60,14 +60,10 @@ export async function main() {
         id: user1.user,
         role: "presenter"
       },
-      {
-        id: user3.user,
-        role: "attendee"
-      }
     ]
-  };console.log
+  };
   const updateParticipants = await roomsClient.updateParticipants(roomId, updateParticipantsRequest);
-  write(`Updated Participants`);
+  console.log(`Updated Participants`);
   printParticipants(updateParticipants);
 
   const deleteUser = {
@@ -88,13 +84,6 @@ export async function main() {
   await roomsClient.deleteRoom(roomId);
 }
 
-function write(message: string): void {
-  const fs = require("fs");
-  fs.writeFileSync("./logs.txt",message,{
-    flag: 'w',
-  });
-}
-
 function printParticipants(pc: ParticipantsCollection): void {
   for (const participant of pc.participants!) {
     const identifierKind = getIdentifierKind(participant.id);
@@ -112,9 +101,13 @@ function printParticipants(pc: ParticipantsCollection): void {
         break;
       case "unknown":
         id = identifierKind.id;
-        write("Unknown user");
+        console.log("Unknown user");
         break;
     }
-    write(`${id} - ${role}`);
+    console.log(`${id} - ${role}`);
   }
 }
+main().catch((error) => {
+  console.error("Encountered an error while sending request: ", error);
+  process.exit(1);
+});
